backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
supabase==2.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# File Processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# AI/LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.7.1
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.20.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Async Processing
celery==5.3.4
redis==5.0.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

